Split pane_group orchestration code into submodules
Move orchestration-specific code out of pane_group/mod.rs into three
new submodules to reduce the size of mod.rs (~9k → ~8k lines).

No behavioral changes — purely a mechanical move.

New file structure under app/src/pane_group/:
  - ambient_pane_restoration.rs (new)
    * AmbientRestoreKind enum
    * register_pending_ambient_restorations
    * process_pending_ambient_restorations
    * replace_pane_with_new_cloud_conversation
    * fetch_and_load_transcript
  - child_agent/hydration.rs (new)
    * RemoteChildHydrationAction enum
    * decide_remote_child_hydration_action fn
    * hydrate_task_backed_hidden_child_pane
    * attempt_remote_child_hydration
    * apply_existing_ambient_task_to_pane
    * hydrate_remote_child_transcript_in_place
    * attach_ambient_session_and_maybe_tombstone
    * process_pending_remote_child_hydrations
  - child_agent/restoration.rs (new)
    * create_hidden_child_agent_pane
    * restore_missing_child_agent_panes_for_parent
    * restore_missing_child_agent_panes_for_terminal_pane_if_needed
    * ensure_hidden_child_agent_pane_for_conversation
    * ensure_shared_session_viewer_child_pane

mod.rs still owns the shared subscription wiring
(ensure_pending_ambient_restoration_subscription and
handle_pending_ambient_restoration_event) plus all PaneGroup struct
fields.
